# project/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Moon Studio API",
        "description": "API dokumentasi untuk sistem booking studio foto Moon Studio",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Moon Studio API Server"
        }
    ],
    "paths": {
...
        "/api/studios": {
            "get": {
                "tags": [
                    "Studios"
                ],
                "summary": "Get all studios",
                "description": "Returns a list of all studios",
                "operationId": "getStudios",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "encrypted string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Studios"
                ],
                "summary": "Create a new studio",
                "description": "Stores a new studio",
                "operationId": "createStudio",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price_per_hour",
                                    "capacity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Studio A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Studio untuk produk dan portrait"
                                    },
                                    "price_per_hour": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250000
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Studio created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "encrypted string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/studios/{id}": {
            "get": {
                "tags": [
                    "Studios"
                ],
                "summary": "Get a single studio",
                "description": "Returns studio detail",
                "operationId": "getStudioById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Studio found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "encrypted string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Studio not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Studios"
                ],
                "summary": "Update studio",
                "description": "Update studio data by ID",
                "operationId": "updateStudio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Studio Baru"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi baru"
                                    },
                                    "price_per_hour": {
                                        "type": "number",
                                        "example": 300000
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Studio updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "encrypted string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Studios"
                ],
                "summary": "Delete studio",
                "description": "Deletes studio by ID",
                "operationId": "deleteStudio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Studio deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "encrypted string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/studios/decrypt": {
            "post": {
                "tags": [
                    "Studios"
                ],
                "summary": "Decrypt encrypted studio data",
                "description": "Returns original data from encrypted studio response",
                "operationId": "decryptStudioData",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "string",
                                        "example": "encrypted-data"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Decryption successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key untuk mengakses endpoint yang dilindungi",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}